refactor(cli): split vscode compatibility layer into separate handlers
